backend/django-project/requirements.txt
Django>=5.2.5
psycopg[binary]>=3.1.19
djangorestframework>=3.15.2
dj-rest-auth[with_social]>=6.0.0
drf-spectacular>=0.27.2
djangorestframework-simplejwt>=5.3.1
django-cors-headers==4.7.0